🛠️: refactor `FormInput` test to typescript
